@@ -5,26 +5,10 @@
  * @package ajaxupload
  * @subpackage lexicon
  */
-$_lang['ajaxupload.cancel'] = 'Abbrechen';
-$_lang['ajaxupload.clearButton'] = 'Alles löschen';
-$_lang['ajaxupload.deleteButton'] = 'Löschen';
-$_lang['ajaxupload.dropArea'] = 'Die Dateien zum Hochladen hier per Drag&amp;Drop einfügen.';
-$_lang['ajaxupload.emptyError'] = '{file} ist leer, bitte wählen sie eine andere Datei.';
-$_lang['ajaxupload.emptyFile'] = 'Die Datei ist leer.';
-$_lang['ajaxupload.failed'] = 'Fehler';
-$_lang['ajaxupload.largeFile'] = 'Die Datei ist zu groß.';
-$_lang['ajaxupload.maxFiles'] = 'Sie dürfen maximal [[+maxFiles]] verschiedene Dateien hochladen.';
-$_lang['ajaxupload.minSizeError'] = '{file} ist zu klein, die minimale Dateigröße beträgt {minSizeLimit}.';
-$_lang['ajaxupload.noFile'] = 'Es wurde keine Datei hochgeladen.';
-$_lang['ajaxupload.notFound'] = 'Die Datei wurde nicht gefunden.';
-$_lang['ajaxupload.notWritable'] = 'Serverfehler. Das Upload-Verzeichis ist nicht beschreibbar.';
-$_lang['ajaxupload.onLeave'] = 'Die Dateien werden gerade hochgeladen. Wenn Sie jetzt die Seite verlassen, wird der Upload abgebrochen.';
-$_lang['ajaxupload.saveError'] = 'Die hochgeladene Datei konnte nicht gespeichert werden. Entweder wurde der Upload abgebrochen oder es ist ein Serverfehler aufgetreten.';
-$_lang['ajaxupload.sizeError'] = '{file} ist zu groß, die maximale Dateigröße beträgt {sizeLimit}.';
+$_lang['ajaxupload'] = 'AjaxUpload';
+$_lang['ajaxupload.cacheNotCreatable'] = 'Der Cache-Ordner konnte nicht erstellt werden!';
+$_lang['ajaxupload.filledFileNotFound'] = 'Die angegebene Datei konnte nicht gefunden werden.';
+$_lang['ajaxupload.missingParameterAjaxuploadTarget'] = 'Fehlender Parameter ajaxuploadTarget.';
+$_lang['ajaxupload.missingParameterAjaxuploadTargetRelativePath'] = 'Fehlender Parameter ajaxuploadTargetRelativePath.';
 $_lang['ajaxupload.targetNotCreatable'] = 'Der Zielordner konnte nicht erstellt werden!';
-$_lang['ajaxupload.targetNotWritable'] = 'Die Datei konnte nicht in das Zielverzeichnis kopiert werden. Existiert das Zielverzeichnis und ist es beschreibbar?';
-$_lang['ajaxupload.thumbnailGenerationProblem'] = 'Das Vorschaubild konnte nicht erzeugt werden.';
-$_lang['ajaxupload.typeError'] = 'Unglücklicherweise entspricht der Dateityp der gewählten Datei(en) nicht den erwarteten Dateitypen. Es sind nur Dateien mit den Erweiterungen {extensions} erlaubt.';
-$_lang['ajaxupload.uploadButton'] = 'Hochladen';
 $_lang['ajaxupload.uploadRequired'] = 'Sie müssen mindestens eine Datei hochladen!';
-$_lang['ajaxupload.wrongExtension'] = 'Die Datei hat eine unerlaubte Erweiterung. Es sind nur Dateien mit den Erweiterungen [[+allowedExtensions]] erlaubt.';
