@@ -5,31 +5,47 @@
  * @package ajaxupload
  * @subpackage lexicon
  */
+$_lang['ajaxupload.ajaxupload.acceptedFileTypes'] = 'Akzeptierte Datei-Mime-Typen für den Upload.';
 $_lang['ajaxupload.ajaxupload.addCss'] = 'CSS am Ende des Head-Bereichs einfügen.';
-$_lang['ajaxupload.ajaxupload.addJquery'] = 'jQuery Javascript am Ende des Body-Bereichs einfügen.';
 $_lang['ajaxupload.ajaxupload.addJscript'] = 'Snippet Javascript und das Fileuploader Script am Ende des Body-Bereichs einfügen.';
-$_lang['ajaxupload.ajaxupload.allowedExtensions'] = 'Erlaubte Erweiterungen für den Upload.';
-$_lang['ajaxupload.ajaxupload.language'] = 'Sprache für das Snippet/Javascript.';
+$_lang['ajaxupload.ajaxupload.fieldformat'] = 'Der Wert der Eigenschaft ajaxuploadFieldformat, der im entsprechenden FormIt-Snippet-Aufruf verwendet wird.';
+$_lang['ajaxupload.ajaxupload.maxFiles'] = 'Maximale Anzahl der hochzuladenden Dateien.';
+$_lang['ajaxupload.ajaxupload.maxFileSize'] = 'Maximale Größe einer hochgeladenen Datei.';
+$_lang['ajaxupload.ajaxupload.placeholderPrefix'] = 'Der Wert der Eigenschaft placeholderPrefix, der im entsprechenden FormIt-Snippet-Aufruf verwendet wird.';
 $_lang['ajaxupload.ajaxupload.maxFiles'] = 'Maximale Anzahl der hochgeladenen Dateien.';
-$_lang['ajaxupload.ajaxupload.maxFilesizeMb'] = 'Maximale Größe einer hochgeladenen Datei.';
 $_lang['ajaxupload.ajaxupload.scriptTpl'] = 'Name eines Chunks, der den Javascript-Code für den Upload-Bereich enthält.';
-$_lang['ajaxupload.ajaxupload.thumbX'] = 'Breite des generierten Vorschaubildes.';
-$_lang['ajaxupload.ajaxupload.thumbY'] = 'Höhe des generierten Vorschaubildes.';
-$_lang['ajaxupload.ajaxupload.uid'] = 'Eindeutige ID für die Upload Queue.';
+$_lang['ajaxupload.ajaxupload.showCredits'] = 'Zeige die Credits von pqina/FilePond javascript.';
+$_lang['ajaxupload.ajaxupload.uid'] = 'Kommagetrennte Liste von eindeutigen IDs für die Upload-Warteschlange.';
 $_lang['ajaxupload.ajaxupload.uploadSectionTpl'] = 'Name eines Chunks, der den HTML-Code für den Upload-Bereich enthält.';
+$_lang['ajaxupload.ajaxupload.value'] = 'Kommagetrennte Liste von Dateien in der Upload-Warteschlange. Ersetzt den Wert, der mit dem Formit2AjaxUpload-Hook abgerufen wird.';
 $_lang['ajaxupload.ajaxupload2formit.ajaxuploadAllowOverwrite'] = 'Überschreiben vorhandener Dateien mit demselben Dateinamen zulassen.';
-$_lang['ajaxupload.ajaxupload2formit.ajaxuploadClearQueue'] = 'Upload Queue nach dem erfolgreichen Ausführen des Hooks leeren.';
-$_lang['ajaxupload.ajaxupload2formit.ajaxuploadFieldformat'] = 'Format der in `ajaxuploadFieldname` gespeicherten Daten.';
-$_lang['ajaxupload.ajaxupload2formit.ajaxuploadFieldname'] = 'Formit Feld, in dem Dateiname und Pfad der hochgeladenen Dateien gespeichert werden.';
+$_lang['ajaxupload.ajaxupload2formit.ajaxuploadCacheExpires'] = 'Verfallszeit des AjaxUpload-Caches (in Stunden).';
+$_lang['ajaxupload.ajaxupload2formit.ajaxuploadFieldformat'] = 'Format der in `ajaxuploadUid` gespeicherten Daten.';
 $_lang['ajaxupload.ajaxupload2formit.ajaxuploadSanitizeFilename'] = 'Bereinigen Sie den Dateinamen der hochgeladenen Datei.';
-$_lang['ajaxupload.ajaxupload2formit.ajaxuploadTarget'] = 'Pfad in dem die hochgeladenen Dateien abgespeichert werden (Relative zo $modx->getOption[\'assetsPath\']). Der Ordner sollte existieren oder erstellbar sein und er muss für PHP beschreibbar sein.';
-$_lang['ajaxupload.ajaxupload2formit.ajaxuploadUid'] = 'Eindeutige ID für die Upload Queue.';
+$_lang['ajaxupload.ajaxupload2formit.ajaxuploadTarget'] = 'Zielpfad für die hochgeladenen Dateien. Der Ordner sollte existieren oder erstellt werden und muss für PHP beschreibbar sein.';
+$_lang['ajaxupload.ajaxupload2formit.ajaxuploadTargetRelativePath'] = 'Der Pfad, mit dem der Zielpfad des Uploads verglichen wird. Die Pfade der hochgeladenen Dateien werden von FormIt relativ zu diesem Pfad verarbeitet.';
+$_lang['ajaxupload.ajaxupload2formit.ajaxuploadUid'] = 'Kommagetrennte Liste von eindeutigen IDs für die Upload-Warteschlange. Außerdem die Namen der in den Hooks verwendeten Felder.';
 $_lang['ajaxupload.ajaxupload2formit.csv'] = 'CSV';
 $_lang['ajaxupload.ajaxupload2formit.json'] = 'JSON';
-$_lang['ajaxupload.ajaxuploadrequired.ajaxuploadRequiredMessage'] = 'Die Fehlermeldung, die hinzugefügt wird, wenn keine Datei hochgeladen wird. Sie entspricht standardmäßig dem Lexikoneintrag `ajaxupload.uploadRequired`.';
-$_lang['ajaxupload.formit2ajaxupload.ajaxuploadFieldformat'] = 'Format der in `ajaxuploadFieldname` gespeicherten Daten.';
-$_lang['ajaxupload.formit2ajaxupload.ajaxuploadFieldname'] = 'Formit Feld, in dem Dateiname und Pfad der hochgeladenen Dateien gespeichert werden.';
-$_lang['ajaxupload.formit2ajaxupload.ajaxuploadTarget'] = 'Pfad in dem die hochgeladenen Dateien abgespeichert werden (Relativ zu $modx->getOption[\'assetsPath\']).';
-$_lang['ajaxupload.formit2ajaxupload.ajaxuploadUid'] = 'Eindeutige ID für die Upload Queue.';
+$_lang['ajaxupload.ajaxuploadattachments.ajaxuploadFieldformat'] = 'Format der in `ajaxuploadUid` gespeicherten Daten.';
+$_lang['ajaxupload.ajaxuploadattachments.ajaxuploadTargetRelativePath'] = 'Der Pfad, mit dem der Zielpfad des Uploads verglichen wird. Die Pfade der hochgeladenen Dateien werden von FormIt relativ zu diesem Pfad verarbeitet.';
+$_lang['ajaxupload.ajaxuploadattachments.ajaxuploadUid'] = 'Kommagetrennte Liste von eindeutigen IDs für die Upload-Warteschlange. Außerdem die Namen der in den Hooks verwendeten Felder.';
+$_lang['ajaxupload.ajaxuploadremove.ajaxuploadFieldformat'] = 'Format der in `ajaxuploadUid` gespeicherten Daten.';
+$_lang['ajaxupload.ajaxuploadremove.ajaxuploadTargetRelativePath'] = 'Der Pfad, mit dem der Zielpfad des Uploads verglichen wird. Die Pfade der hochgeladenen Dateien werden von FormIt relativ zu diesem Pfad verarbeitet.';
+$_lang['ajaxupload.ajaxuploadremove.ajaxuploadUid'] = 'Kommagetrennte Liste von eindeutigen IDs für die Upload-Warteschlange. Außerdem die Namen der in den Hooks verwendeten Felder.';
+$_lang['ajaxupload.ajaxuploadrequired.ajaxuploadFieldformat'] = 'Format der in `ajaxuploadUid` gespeicherten Daten.';
+$_lang['ajaxupload.ajaxuploadrequired.ajaxuploadRequiredMessage'] = 'Die Fehlermeldung, die hinzugefügt wird, wenn keine Datei hochgeladen wird.';
+$_lang['ajaxupload.ajaxuploadrequired.ajaxuploadUid'] = 'Kommagetrennte Liste von eindeutigen IDs für die Upload-Warteschlange. Außerdem die Namen der in den Hooks verwendeten Felder.';
+$_lang['ajaxupload.formit2ajaxupload.ajaxuploadCacheExpires'] = 'Verfallszeit des AjaxUpload-Caches (in Stunden).';
+$_lang['ajaxupload.formit2ajaxupload.ajaxuploadFieldformat'] = 'Format der in `ajaxuploadUid` gespeicherten Daten.';
+$_lang['ajaxupload.formit2ajaxupload.ajaxuploadTarget'] = 'Zielpfad, der die vorhandenen Dateien enthält, die für das FilePond-Javascript vorbereitet werden.';
+$_lang['ajaxupload.formit2ajaxupload.ajaxuploadTargetRelativePath'] = 'Der Pfad, mit dem der Zielpfad des Uploads verglichen wird. Die Pfade der hochgeladenen Dateien werden von FormIt relativ zu diesem Pfad verarbeitet.';
+$_lang['ajaxupload.formit2ajaxupload.ajaxuploadUid'] = 'Kommagetrennte Liste von eindeutigen IDs für die Upload-Warteschlange. Außerdem die Namen der in den Hooks verwendeten Felder.';
 $_lang['ajaxupload.formit2ajaxupload.csv'] = 'CSV';
 $_lang['ajaxupload.formit2ajaxupload.json'] = 'JSON';
+$_lang['ajaxupload.ajaxupload.targetRelativePath'] = 'Der Wert der Eigenschaft ajaxuploadTargetRelativePath, der im entsprechenden FormIt-Snippet-Aufruf verwendet wird.';
+$_lang['ajaxupload.ajaxuploadremove.ajaxuploadTargetMediasource'] = 'Wenn nicht leer, wird dem Zielpfad der Base-Pfad der Medienquelle vorangestellt. Andernfalls wird dem Zielpfad der MODX-Assets-Pfad vorangestellt.';
+$_lang['ajaxupload.formit2ajaxupload.ajaxuploadTargetMediasource'] = 'Wenn nicht leer, wird dem Zielpfad der Base-Pfad der Medienquelle vorangestellt. Andernfalls wird dem Zielpfad der MODX-Assets-Pfad vorangestellt.';
+$_lang['ajaxupload.ajaxupload.targetMediasource'] = 'Der Wert der Eigenschaft ajaxuploadTargetMediasource, der im entsprechenden FormIt-Snippet-Aufruf verwendet wird.';
+$_lang['ajaxupload.ajaxupload2formit.ajaxuploadTargetMediasource'] = 'Wenn nicht leer, wird dem Zielpfad der Base-Pfad der Medienquelle vorangestellt. Andernfalls wird dem Zielpfad der MODX-Assets-Pfad vorangestellt.';
+$_lang['ajaxupload.ajaxuploadattachments.ajaxuploadTargetMediasource'] = 'Wenn nicht leer, wird dem Zielpfad der Base-Pfad der Medienquelle vorangestellt. Andernfalls wird dem Zielpfad der MODX-Assets-Pfad vorangestellt.';
