@@ -5,26 +5,9 @@
  * @package ajaxupload
  * @subpackage lexicon
  */
-$_lang['ajaxupload.cancel'] = 'Cancel';
-$_lang['ajaxupload.clearButton'] = 'Delete all';
-$_lang['ajaxupload.deleteButton'] = 'Delete';
-$_lang['ajaxupload.dropArea'] = 'Drop files here to upload.';
-$_lang['ajaxupload.emptyError'] = '{file} is empty, please select files again without it.';
-$_lang['ajaxupload.emptyFile'] = 'File is empty.';
-$_lang['ajaxupload.failed'] = 'Failed';
-$_lang['ajaxupload.largeFile'] = 'File is too large.';
-$_lang['ajaxupload.maxFiles'] = 'Only [[+maxFiles]] different files are allowed to upload.';
-$_lang['ajaxupload.minSizeError'] = '{file} is too small, minimum file size is {minSizeLimit}.';
-$_lang['ajaxupload.noFile'] = 'No files were uploaded.';
-$_lang['ajaxupload.notFound'] = 'File is not found.';
-$_lang['ajaxupload.notWritable'] = 'Server error. Upload directory isn’t writable.';
-$_lang['ajaxupload.onLeave'] = 'The files are being uploaded, if you leave now the upload will be cancelled.';
-$_lang['ajaxupload.saveError'] = 'Could not save uploaded file. The upload was cancelled or a server error encountered.';
-$_lang['ajaxupload.sizeError'] = '{file} is too large, maximum file size is {sizeLimit}.';
+$_lang['ajaxupload'] = 'AjaxUpload';
+$_lang['ajaxupload.cacheNotCreatable'] = 'Could not create the cache folder!';
+$_lang['ajaxupload.filledFileNotFound'] = 'The specified file could not be found.';
+$_lang['ajaxupload.missingParameterAjaxuploadTarget'] = 'Missing parameter ajaxuploadTarget.';
 $_lang['ajaxupload.targetNotCreatable'] = 'Could not create the target folder!';
-$_lang['ajaxupload.targetNotWritable'] = 'Could not copy the uploaded file to the target folder. Does the folder exist and is it writable?';
-$_lang['ajaxupload.thumbnailGenerationProblem'] = 'Thumbnail generation failed.';
-$_lang['ajaxupload.typeError'] = 'Unfortunately the file(s) you selected weren’t the type we were expecting. Only {extensions} files are allowed.';
-$_lang['ajaxupload.uploadButton'] = 'Upload';
 $_lang['ajaxupload.uploadRequired'] = 'You have to upload at least one file!';
-$_lang['ajaxupload.wrongExtension'] = 'File has an invalid extension, it should be one of [[+allowedExtensions]].';
