@@ -5,31 +5,41 @@
  * @package ajaxupload
  * @subpackage lexicon
  */
+$_lang['ajaxupload.ajaxupload.acceptedFileTypes'] = 'Accepted file mime types for upload.';
 $_lang['ajaxupload.ajaxupload.addCss'] = 'Add the CSS at the end of the head.';
-$_lang['ajaxupload.ajaxupload.addJquery'] = 'Add jQuery script at the end of the body.';
 $_lang['ajaxupload.ajaxupload.addJscript'] = 'Add the snippet javascript and the fileuploader script at the end of the body.';
-$_lang['ajaxupload.ajaxupload.allowedExtensions'] = 'Allowed file extensions for upload.';
-$_lang['ajaxupload.ajaxupload.language'] = 'Snippet/Javascript language.';
+$_lang['ajaxupload.ajaxupload.fieldformat'] = 'The value of ajaxuploadFieldformat property used in the according FormIt snippet call.';
 $_lang['ajaxupload.ajaxupload.maxFiles'] = 'Maximum count of files to upload.';
-$_lang['ajaxupload.ajaxupload.maxFilesizeMb'] = 'Maximum size for one file to upload.';
+$_lang['ajaxupload.ajaxupload.maxFileSize'] = 'Maximum size for one file to upload.';
+$_lang['ajaxupload.ajaxupload.placeholderPrefix'] = 'The value of placeholderPrefix property used in the according FormIt snippet call.';
 $_lang['ajaxupload.ajaxupload.scriptTpl'] = 'Name of a chunk that contains the Javascript code for the upload section.';
-$_lang['ajaxupload.ajaxupload.thumbX'] = 'Horizontal size of generated thumb.';
-$_lang['ajaxupload.ajaxupload.thumbY'] = 'Vertical size of generated thumb.';
-$_lang['ajaxupload.ajaxupload.uid'] = 'Unique upload queue id.';
+$_lang['ajaxupload.ajaxupload.showCredits'] = 'Show the credits of pqina/FilePond javascript.';
+$_lang['ajaxupload.ajaxupload.targetMediasource'] = 'The value of ajaxuploadTargetMediasource property used in the according FormIt snippet call.';
+$_lang['ajaxupload.ajaxupload.uid'] = 'Comma separated list of unique upload queue ids.';
 $_lang['ajaxupload.ajaxupload.uploadSectionTpl'] = 'Name of a chunk that contains the HTML code for the upload section.';
+$_lang['ajaxupload.ajaxupload.value'] = 'Comma separated list of files in the upload queue. Supercedes the value that is retrieved with the Formit2AjaxUpload hook.';
 $_lang['ajaxupload.ajaxupload2formit.ajaxuploadAllowOverwrite'] = 'Allow overwrite of existing files with the same filename.';
-$_lang['ajaxupload.ajaxupload2formit.ajaxuploadClearQueue'] = 'Clear the upload queue after a successful run of the hook.';
-$_lang['ajaxupload.ajaxupload2formit.ajaxuploadFieldformat'] = 'Format of the data saved in `ajaxuploadFieldname`.';
-$_lang['ajaxupload.ajaxupload2formit.ajaxuploadFieldname'] = 'Formit field, the filenames/paths of the (already) uploaded files are saved in.';
+$_lang['ajaxupload.ajaxupload2formit.ajaxuploadCacheExpires'] = 'Expire Time of the AjaxUpload cache (in hours).';
+$_lang['ajaxupload.ajaxupload2formit.ajaxuploadFieldformat'] = 'Format of the data saved in `ajaxuploadUid`.';
 $_lang['ajaxupload.ajaxupload2formit.ajaxuploadSanitizeFilename'] = 'Sanitize the filename of the uploaded file.';
-$_lang['ajaxupload.ajaxupload2formit.ajaxuploadTarget'] = 'Target path for the (already) uploaded files (relative to $modx->getOption[\'assetsPath\']). The folder should exist or should be createable and it has to be writable for PHP.';
-$_lang['ajaxupload.ajaxupload2formit.ajaxuploadUid'] = 'Unique upload queue id.';
+$_lang['ajaxupload.ajaxupload2formit.ajaxuploadTarget'] = 'Target path for the uploaded files. The folder should exist or be created and must be writable for PHP.';
+$_lang['ajaxupload.ajaxupload2formit.ajaxuploadTargetMediasource'] = 'If not empty, The target path is prefixed with the media source base path. Otherwise the target path is prefixed with the MODX assets path.';
+$_lang['ajaxupload.ajaxupload2formit.ajaxuploadUid'] = 'Comma separated list of unique upload queue ids. Also the names of the fields used in the hooks.';
 $_lang['ajaxupload.ajaxupload2formit.csv'] = 'CSV';
 $_lang['ajaxupload.ajaxupload2formit.json'] = 'JSON';
-$_lang['ajaxupload.ajaxuploadrequired.ajaxuploadRequiredMessage'] = 'The error message added, when no file is uploaded. It defaults to the lexicon entry `ajaxupload.uploadRequired`.';
-$_lang['ajaxupload.formit2ajaxupload.ajaxuploadFieldformat'] = 'Format of the data saved in `ajaxuploadFieldname`.';
-$_lang['ajaxupload.formit2ajaxupload.ajaxuploadFieldname'] = 'Formit field, the filenames/paths of the (already) uploaded files are saved in.';
-$_lang['ajaxupload.formit2ajaxupload.ajaxuploadTarget'] = 'Target path for the (already) uploaded files (relative to $modx->getOption[\'assetsPath\']).';
-$_lang['ajaxupload.formit2ajaxupload.ajaxuploadUid'] = 'Unique upload queue id.';
+$_lang['ajaxupload.ajaxuploadattachments.ajaxuploadFieldformat'] = 'Format of the data saved in `ajaxuploadUid`.';
+$_lang['ajaxupload.ajaxuploadattachments.ajaxuploadTargetMediasource'] = 'If not empty, The target path is prefixed with the media source base path. Otherwise the target path is prefixed with the MODX assets path.';
+$_lang['ajaxupload.ajaxuploadattachments.ajaxuploadUid'] = 'Comma separated list of unique upload queue ids. Also the names of the fields used in the hooks.';
+$_lang['ajaxupload.ajaxuploadremove.ajaxuploadFieldformat'] = 'Format of the data saved in `ajaxuploadUid`.';
+$_lang['ajaxupload.ajaxuploadremove.ajaxuploadTargetMediasource'] = 'If not empty, The target path is prefixed with the media source base path. Otherwise the target path is prefixed with the MODX assets path.';
+$_lang['ajaxupload.ajaxuploadremove.ajaxuploadUid'] = 'Comma separated list of unique upload queue ids. Also the names of the fields used in the hooks.';
+$_lang['ajaxupload.ajaxuploadrequired.ajaxuploadFieldformat'] = 'Format of the data saved in `ajaxuploadUid`.';
+$_lang['ajaxupload.ajaxuploadrequired.ajaxuploadRequiredMessage'] = 'The error message added, when no file is uploaded.';
+$_lang['ajaxupload.ajaxuploadrequired.ajaxuploadUid'] = 'Comma separated list of unique upload queue ids. Also the names of the fields used in the hooks.';
+$_lang['ajaxupload.formit2ajaxupload.ajaxuploadCacheExpires'] = 'Expire Time of the AjaxUpload cache (in hours).';
+$_lang['ajaxupload.formit2ajaxupload.ajaxuploadFieldformat'] = 'Format of the data saved in `ajaxuploadUid`.';
+$_lang['ajaxupload.formit2ajaxupload.ajaxuploadTarget'] = 'Target path that contains the existing files, that are prepared for the FilePond javascript.';
+$_lang['ajaxupload.formit2ajaxupload.ajaxuploadTargetMediasource'] = 'If not empty, The target path is prefixed with the media source base path. Otherwise the target path is prefixed with the MODX assets path.';
+$_lang['ajaxupload.formit2ajaxupload.ajaxuploadUid'] = 'Comma separated list of unique upload queue ids. Also the names of the fields used in the hooks.';
 $_lang['ajaxupload.formit2ajaxupload.csv'] = 'CSV';
 $_lang['ajaxupload.formit2ajaxupload.json'] = 'JSON';
