@@ -5,26 +5,10 @@
  * @package ajaxupload
  * @subpackage lexicon
  */
-$_lang['ajaxupload.cancel'] = 'Annuleren';
-$_lang['ajaxupload.clearButton'] = 'Alles verwijderen';
-$_lang['ajaxupload.deleteButton'] = 'Verwijderen';
-$_lang['ajaxupload.dropArea'] = 'Sleep bestanden hier naartoe om te uploaden.';
-$_lang['ajaxupload.emptyError'] = '{file} is leeg, probeer het opnieuw of selecteer een ander bestand.';
-$_lang['ajaxupload.emptyFile'] = 'Bestand is leeg.';
-$_lang['ajaxupload.failed'] = 'Mislukt';
-$_lang['ajaxupload.largeFile'] = 'Het bestand is te groot.';
-$_lang['ajaxupload.maxFiles'] = 'Er mogen slechts [[+maxFiles]] verschillende bestanden worden geüpload.';
-$_lang['ajaxupload.minSizeError'] = '{file} is te klein, minimale bestandsgrootte is {minSizeLimit}.';
-$_lang['ajaxupload.noFile'] = 'Er zijn geen bestanden geüpload.';
-$_lang['ajaxupload.notFound'] = 'Bestand niet gevonden.';
-$_lang['ajaxupload.notWritable'] = 'Serverfout. Geen schrijfrechten in upload map.';
-$_lang['ajaxupload.onLeave'] = 'De bestanden zijn nog onderweg. Als je nu weggaat, wordt de upload geannuleerd.';
-$_lang['ajaxupload.saveError'] = 'Kon het geüploade bestand niet opslaan. De upload is geannuleerd of er is een serverfout opgetreden.';
-$_lang['ajaxupload.sizeError'] = '{file} is te groot, maximale bestandsgrootte is {sizeLimit}.';
-$_lang['ajaxupload.targetNotCreatable'] = 'Kan de doelmap niet aanmaken!';
-$_lang['ajaxupload.targetNotWritable'] = 'Kan het geüploade bestand niet kopiëren naar de doelmap. Bestaat de map en is deze beschrijfbaar?';
-$_lang['ajaxupload.thumbnailGenerationProblem'] = 'Thumbnail aanmaken mislukt.';
-$_lang['ajaxupload.typeError'] = 'Er zitten ongeldige bestandsformaten tussen de geselecteerde bestanden. Alleen de volgende bestanden zijn toegestaan: {extensies}.';
-$_lang['ajaxupload.uploadButton'] = 'Upload';
-$_lang['ajaxupload.uploadRequired'] = 'Je moet minstens één bestand uploaden!';
-$_lang['ajaxupload.wrongExtension'] = 'Bestand heeft een ongeldige extensie. Geaccepteerde extensies: [[+allowedExtensions]].';
+$_lang['ajaxupload'] = 'AjaxUpload';
+$_lang['ajaxupload.cacheNotCreatable'] = 'Kan de cache-map niet maken!';
+$_lang['ajaxupload.filledFileNotFound'] = 'Het opgegeven bestand kon niet worden gevonden.';
+$_lang['ajaxupload.missingParameterAjaxuploadTarget'] = 'Parameter ajaxuploadTarget ontbreekt.';
+$_lang['ajaxupload.missingParameterAjaxuploadTargetRelativePath'] = 'Parameter ajaxuploadTargetRelativePath ontbreekt.';
+$_lang['ajaxupload.targetNotCreatable'] = 'Kan de doelmap niet maken!';
+$_lang['ajaxupload.uploadRequired'] = 'U moet ten minste één bestand uploaden!';
