@@ -5,31 +5,46 @@
  * @package ajaxupload
  * @subpackage lexicon
  */
+$_lang['ajaxupload.ajaxupload.acceptedFileTypes'] = 'Geaccepteerde bestandstypes voor upload.';
 $_lang['ajaxupload.ajaxupload.addCss'] = 'Voeg de CSS toe aan het eind van de head.';
-$_lang['ajaxupload.ajaxupload.addJquery'] = 'Voeg jQuery script toe aan het eind van de body.';
 $_lang['ajaxupload.ajaxupload.addJscript'] = 'Voeg de Javascript snippet en het fileuploader script toe aan het eind van de body.';
-$_lang['ajaxupload.ajaxupload.allowedExtensions'] = 'Deze bestandsextensies zijn toegestaan.';
-$_lang['ajaxupload.ajaxupload.language'] = 'Snippet/Javascript taal.';
+$_lang['ajaxupload.ajaxupload.fieldformat'] = 'De waarde van de eigenschap ajaxuploadFieldformat die wordt gebruikt in de bijbehorende FormIt-snippetoproep.';
 $_lang['ajaxupload.ajaxupload.maxFiles'] = 'Maximaal aantal bestanden voor upload.';
-$_lang['ajaxupload.ajaxupload.maxFilesizeMb'] = 'Maximum bestandsgrootte per bestand.';
+$_lang['ajaxupload.ajaxupload.maxFileSize'] = 'Maximum bestandsgrootte per bestand.';
+$_lang['ajaxupload.ajaxupload.placeholderPrefix'] = 'De waarde van de eigenschap placeholderPrefix die wordt gebruikt in de bijbehorende FormIt-snippetoproep.';
 $_lang['ajaxupload.ajaxupload.scriptTpl'] = 'Naam van een chunk waar de Javascript code voor het upload gedeelte in staat.';
-$_lang['ajaxupload.ajaxupload.thumbX'] = 'Breedte van de gegenereerde thumbnail.';
-$_lang['ajaxupload.ajaxupload.thumbY'] = 'Hoogte van de gegenereerde thumbnail.';
-$_lang['ajaxupload.ajaxupload.uid'] = 'Uniek ID voor upload wachtrij.';
+$_lang['ajaxupload.ajaxupload.showCredits'] = 'Toon de credits van pqina/FilePond javascript.';
+$_lang['ajaxupload.ajaxupload.uid'] = 'Door komma’s gescheiden lijst van unieke uploadwachtrij-id’s.';
 $_lang['ajaxupload.ajaxupload.uploadSectionTpl'] = 'Naam van een chunk waar de HTML code voor het upload gedeelte in staat.';
+$_lang['ajaxupload.ajaxupload.value'] = 'Door komma’s gescheiden lijst van bestanden in de uploadwachtrij. Vervangt de waarde die wordt opgehaald met de Formit2AjaxUpload haak.';
 $_lang['ajaxupload.ajaxupload2formit.ajaxuploadAllowOverwrite'] = 'Bestaande bestanden met dezelfde naam mogen overschreven worden.';
-$_lang['ajaxupload.ajaxupload2formit.ajaxuploadClearQueue'] = 'Leeg de wachtrij nadat de hook succesvol heeft gedraaid.';
-$_lang['ajaxupload.ajaxupload2formit.ajaxuploadFieldformat'] = 'Formaat van de data opgeslagen in `ajaxuploadFieldname`.';
-$_lang['ajaxupload.ajaxupload2formit.ajaxuploadFieldname'] = 'Het Formit veld waarin de paden/bestandsnamen van reeds voltooide uploads worden opgeslagen.';
+$_lang['ajaxupload.ajaxupload2formit.ajaxuploadCacheExpires'] = 'Verlooptijd van de AjaxUpload-cache (in uren).';
+$_lang['ajaxupload.ajaxupload2formit.ajaxuploadFieldformat'] = 'Formaat van de data opgeslagen in `ajaxuploadUid`.';
 $_lang['ajaxupload.ajaxupload2formit.ajaxuploadSanitizeFilename'] = 'De bestandsnaam van het geüploade bestand zuiveren.';
-$_lang['ajaxupload.ajaxupload2formit.ajaxuploadTarget'] = 'Doel locatie voor reeds geuploade bestanden (relatief aan $modx->getOption[\'assetsPath\']). Zorg ervoor dat PHP schrijfrechten heeft op deze locatie.';
-$_lang['ajaxupload.ajaxupload2formit.ajaxuploadUid'] = 'Uniek ID voor upload wachtrij.';
+$_lang['ajaxupload.ajaxupload2formit.ajaxuploadTarget'] = 'Doelpad voor de geüploade bestanden. De map moet bestaan of aangemaakt zijn en moet schrijfbaar zijn voor PHP.';
+$_lang['ajaxupload.ajaxupload2formit.ajaxuploadTargetRelativePath'] = 'Het pad waarmee het doelpad van de upload wordt vergeleken. De paden van de geüploade bestanden worden door FormIt relatief ten opzichte van dit pad verwerkt.';
+$_lang['ajaxupload.ajaxupload2formit.ajaxuploadUid'] = 'Kommagescheiden lijst van unieke uploadwachtrij-id’s. Ook de namen van de velden die in de haken worden gebruikt.';
 $_lang['ajaxupload.ajaxupload2formit.csv'] = 'CSV';
 $_lang['ajaxupload.ajaxupload2formit.json'] = 'JSON';
-$_lang['ajaxupload.ajaxuploadrequired.ajaxuploadRequiredMessage'] = 'De foutmelding wanneer er niets is upgeload. Valt terug op de lexicon sleutel `ajaxupload.uploadRequired`.';
-$_lang['ajaxupload.formit2ajaxupload.ajaxuploadFieldformat'] = 'Formaat van de data in `ajaxuploadFieldname`.';
-$_lang['ajaxupload.formit2ajaxupload.ajaxuploadFieldname'] = 'Het Formit veld waarin de paden/bestandsnamen van reeds voltooide uploads worden opgeslagen.';
-$_lang['ajaxupload.formit2ajaxupload.ajaxuploadTarget'] = 'Doel locatie voor reeds geuploade bestanden (relatief aan $modx->getOption[\'assetsPath\']).';
-$_lang['ajaxupload.formit2ajaxupload.ajaxuploadUid'] = 'Uniek ID voor upload wachtrij.';
+$_lang['ajaxupload.ajaxuploadattachments.ajaxuploadFieldformat'] = 'Formaat van de gegevens opgeslagen in `ajaxuploadUid`.';
+$_lang['ajaxupload.ajaxuploadattachments.ajaxuploadTargetRelativePath'] = 'Het pad waarmee het doelpad van de upload wordt vergeleken. De paden van de geüploade bestanden worden door FormIt relatief ten opzichte van dit pad verwerkt.';
+$_lang['ajaxupload.ajaxuploadattachments.ajaxuploadUid'] = 'Kommagescheiden lijst van unieke uploadwachtrij-id’s. Ook de namen van de velden die in de haken worden gebruikt.';
+$_lang['ajaxupload.ajaxuploadremove.ajaxuploadFieldformat'] = 'Formaat van de gegevens die zijn opgeslagen in `ajaxuploadUid`.';
+$_lang['ajaxupload.ajaxuploadremove.ajaxuploadTargetRelativePath'] = 'Het pad waarmee het doelpad van de upload wordt vergeleken. De paden van de geüploade bestanden worden door FormIt relatief ten opzichte van dit pad verwerkt.';
+$_lang['ajaxupload.ajaxuploadremove.ajaxuploadUid'] = 'Kommagescheiden lijst van unieke uploadwachtrij-id’s. Ook de namen van de velden die in de haken worden gebruikt.';
+$_lang['ajaxupload.ajaxuploadrequired.ajaxuploadFieldformat'] = 'Formaat van de gegevens die zijn opgeslagen in `ajaxuploadUid`.';
+$_lang['ajaxupload.ajaxuploadrequired.ajaxuploadRequiredMessage'] = 'De foutmelding die wordt toegevoegd als er geen bestand is geüpload.';
+$_lang['ajaxupload.ajaxuploadrequired.ajaxuploadUid'] = 'Kommagescheiden lijst van unieke uploadwachtrij-id’s. Ook de namen van de velden die in de haken worden gebruikt.';
+$_lang['ajaxupload.formit2ajaxupload.ajaxuploadCacheExpires'] = 'Verlooptijd van de AjaxUpload-cache (in uren).';
+$_lang['ajaxupload.formit2ajaxupload.ajaxuploadFieldformat'] = 'Formaat van de data in `ajaxuploadUid`.';
+$_lang['ajaxupload.formit2ajaxupload.ajaxuploadTarget'] = 'Doelpad dat de bestaande bestanden bevat die zijn voorbereid voor het FilePond javascript.';
+$_lang['ajaxupload.formit2ajaxupload.ajaxuploadTargetRelativePath'] = 'Het pad waarmee het doelpad van de upload wordt vergeleken. De paden van de geüploade bestanden worden door FormIt relatief ten opzichte van dit pad verwerkt.';
+$_lang['ajaxupload.formit2ajaxupload.ajaxuploadUid'] = 'Kommagescheiden lijst van unieke uploadwachtrij-id’s. Ook de namen van de velden die in de haken worden gebruikt.';
 $_lang['ajaxupload.formit2ajaxupload.csv'] = 'CSV';
 $_lang['ajaxupload.formit2ajaxupload.json'] = 'JSON';
+$_lang['ajaxupload.ajaxupload.targetRelativePath'] = 'De waarde van de eigenschap ajaxuploadTargetRelativePath die wordt gebruikt in de bijbehorende FormIt-nippetaanroep.';
+$_lang['ajaxupload.ajaxupload2formit.ajaxuploadTargetMediasource'] = 'Als deze niet leeg is, wordt het doelpad voorafgegaan door het basispad van de mediabron. Anders wordt het doelpad voorafgegaan door het pad van MODX-activa.';
+$_lang['ajaxupload.formit2ajaxupload.ajaxuploadTargetMediasource'] = 'Als deze niet leeg is, wordt het doelpad voorafgegaan door het basispad van de mediabron. Anders wordt het doelpad voorafgegaan door het pad van MODX-activa.';
+$_lang['ajaxupload.ajaxuploadremove.ajaxuploadTargetMediasource'] = 'Als deze niet leeg is, wordt het doelpad voorafgegaan door het basispad van de mediabron. Anders wordt het doelpad voorafgegaan door het pad van MODX-activa.';
+$_lang['ajaxupload.ajaxupload.targetMediasource'] = 'De waarde van de eigenschap ajaxuploadTargetMediasource die wordt gebruikt in de bijbehorende FormIt-nippetaanroep.';
+$_lang['ajaxupload.ajaxuploadattachments.ajaxuploadTargetMediasource'] = 'Als deze niet leeg is, wordt het doelpad voorafgegaan door het basispad van de mediabron. Anders wordt het doelpad voorafgegaan door het pad van MODX-activa.';
