@@ -9,11 +9,11 @@ $_lang['setting_ajaxupload.cache_expires'] = 'Vervaltijd';
 $_lang['setting_ajaxupload.cache_expires_desc'] = 'Vervaltijd van de AjaxUpload cache (in uren)';
 $_lang['setting_ajaxupload.debug'] = 'Debug';
 $_lang['setting_ajaxupload.debug_desc'] = 'Log debug-informatie in het MODX log bestand.';
+$_lang['setting_ajaxupload.file_tpl'] = 'Bestandssjabloon';
+$_lang['setting_ajaxupload.file_tpl_desc'] = 'Naam van een brok die de HTML-code bevat voor het weergeven van het geüploade bestand.';
 $_lang['setting_ajaxupload.filename_restrict_chars'] = 'Methode voor beperking van bestandsnamen';
 $_lang['setting_ajaxupload.filename_restrict_chars_desc'] = 'De methode die gebruikt wordt om karakters in de bestandsnaam te beperken. "pattern" laat een RegEx patroon toe, "legal" laat alle legale URL karakters toe, "alpha" laat alleen letters van het alfabet toe en "alphanumeric" laat alleen letters en cijfers toe. Als deze leeg is, wordt de waarde van de equivalente MODX FURL systeeminstelling overgenomen.';
 $_lang['setting_ajaxupload.filename_restrict_chars_pattern'] = 'Karakterbeperkingspatroon bestandsnaam';
 $_lang['setting_ajaxupload.filename_restrict_chars_pattern_desc'] = 'Een geldige RegEx voor het beperken van tekens in de bestandsnaam. Als deze leeg is, wordt de waarde van de equivalente MODX FURL systeeminstelling geërfd.';
 $_lang['setting_ajaxupload.filename_translit'] = 'Bestandsnaam transliteratie methode';
 $_lang['setting_ajaxupload.filename_translit_desc'] = 'Staat standaard op iconv_ascii, om mogelijke problemen met tekens met accenten te voorkomen. Als deze leeg is, wordt de waarde van de equivalente MODX FURL systeeminstelling geërfd.';
-$_lang['setting_ajaxupload.image_tpl'] = 'Template voor afbeelding';
-$_lang['setting_ajaxupload.image_tpl_desc'] = 'Naam van een chunk die de HTML-code bevat om de geüploade afbeelding weer te geven.';
