@@ -5,26 +5,10 @@
  * @package ajaxupload
  * @subpackage lexicon
  */
-$_lang['ajaxupload.cancel'] = 'Отмена';
-$_lang['ajaxupload.clearButton'] = 'Удалить все';
-$_lang['ajaxupload.deleteButton'] = 'Удалить';
-$_lang['ajaxupload.dropArea'] = 'Перетащите файлы сюда.';
-$_lang['ajaxupload.emptyError'] = '{file} пуст. Пожалуйста, выберите файлы без него.';
-$_lang['ajaxupload.emptyFile'] = 'Файл пуст.';
-$_lang['ajaxupload.failed'] = 'Не вышло';
-$_lang['ajaxupload.largeFile'] = 'Файл слишком велик.';
-$_lang['ajaxupload.maxFiles'] = 'Только [[+maxFiles]] файлов можно загрузить.';
-$_lang['ajaxupload.minSizeError'] = '{file} слишком мал, минимальный размер — {minSizeLimit}.';
-$_lang['ajaxupload.noFile'] = 'Файлы не загружены.';
-$_lang['ajaxupload.notFound'] = 'Файл не найден.';
-$_lang['ajaxupload.notWritable'] = 'Ошибка сервера. Каталог загрузки закрыт для записи.';
-$_lang['ajaxupload.onLeave'] = 'Файлы загружаются, если Вы закроете страницу, загрузка будет отменена.';
-$_lang['ajaxupload.saveError'] = 'Загруженный файл не сохранён. Загрузка отменена или произошла ошибка сервера.';
-$_lang['ajaxupload.sizeError'] = '{file} слишком велик, максимальный размер — {sizeLimit}.';
+$_lang['ajaxupload'] = 'AjaxUpload';
+$_lang['ajaxupload.cacheNotCreatable'] = 'Не удалось создать папку кэша!';
+$_lang['ajaxupload.filledFileNotFound'] = 'Указанный файл не найден.';
+$_lang['ajaxupload.missingParameterAjaxuploadTarget'] = 'Отсутствует параметр ajaxuploadTarget.';
+$_lang['ajaxupload.missingParameterAjaxuploadTargetRelativePath'] = 'Отсутствует параметр ajaxuploadTargetRelativePath.';
 $_lang['ajaxupload.targetNotCreatable'] = 'Не удалось создать целевую папку!';
-$_lang['ajaxupload.targetNotWritable'] = 'Не удалось скопировать загруженный файл в целевую папку. Папка существует и доступен для записи?';
-$_lang['ajaxupload.thumbnailGenerationProblem'] = 'Эскиз не создан.';
-$_lang['ajaxupload.typeError'] = 'К сожалению файл(ы) этого типа загрузить нельзя. Только файлы {extensions}.';
-$_lang['ajaxupload.uploadButton'] = 'Прикрепить файл';
 $_lang['ajaxupload.uploadRequired'] = 'Вы должны загрузить хотя бы один файл!';
-$_lang['ajaxupload.wrongExtension'] = 'У файла неверное расширение, он может быть только [[+allowedExtensions]].';
