@@ -5,31 +5,46 @@
  * @package ajaxupload
  * @subpackage lexicon
  */
+$_lang['ajaxupload.ajaxupload.acceptedFileTypes'] = 'Принимаемые типы мимов файлов для загрузки.';
 $_lang['ajaxupload.ajaxupload.addCss'] = 'Добавьте CSS в конец заголовка.';
-$_lang['ajaxupload.ajaxupload.addJquery'] = 'Добавить вызов jQuery в конец body.';
 $_lang['ajaxupload.ajaxupload.addJscript'] = 'Добавить javascript сниппета и вызов скрипта fileuploader script в конец body.';
-$_lang['ajaxupload.ajaxupload.allowedExtensions'] = 'Разрешённые расширения для загрузки.';
-$_lang['ajaxupload.ajaxupload.language'] = 'Язык сниппета/Javascript-а.';
+$_lang['ajaxupload.ajaxupload.fieldformat'] = 'Значение свойства ajaxuploadFieldformat, используемое в соответствующем вызове сниппета FormIt.';
 $_lang['ajaxupload.ajaxupload.maxFiles'] = 'Максимальное число файлов для загрузки.';
-$_lang['ajaxupload.ajaxupload.maxFilesizeMb'] = 'Максимальный размер одного заргужаемого файла.';
+$_lang['ajaxupload.ajaxupload.maxFileSize'] = 'Максимальный размер одного заргужаемого файла.';
+$_lang['ajaxupload.ajaxupload.placeholderPrefix'] = 'Значение свойства placeholderPrefix, используемое в соответствующем вызове сниппета FormIt.';
 $_lang['ajaxupload.ajaxupload.scriptTpl'] = 'Имя чанка, содержащего Javascript-код для раздела загрузки.';
-$_lang['ajaxupload.ajaxupload.thumbX'] = 'Горизонтальный размер генерируемого эскиза.';
-$_lang['ajaxupload.ajaxupload.thumbY'] = 'Вертикальный размер генерируемого эскиза.';
-$_lang['ajaxupload.ajaxupload.uid'] = 'Уникальный id очереди загрузки.';
+$_lang['ajaxupload.ajaxupload.showCredits'] = 'Показать титры pqina/FilePond javascript.';
+$_lang['ajaxupload.ajaxupload.uid'] = 'Список уникальных идентификаторов очередей загрузки, разделенных запятыми.';
 $_lang['ajaxupload.ajaxupload.uploadSectionTpl'] = 'Имя фрагмента, содержащего HTML-код для раздела загрузки.';
+$_lang['ajaxupload.ajaxupload.value'] = 'Список файлов в очереди на загрузку, разделенный запятыми. Заменяет значение, получаемое с помощью хука Formit2AjaxUpload.';
 $_lang['ajaxupload.ajaxupload2formit.ajaxuploadAllowOverwrite'] = 'Разрешить перезапись существующих файлов с тем же именем.';
-$_lang['ajaxupload.ajaxupload2formit.ajaxuploadClearQueue'] = 'Очистите очередь загрузки после успешного выполнения хука.';
-$_lang['ajaxupload.ajaxupload2formit.ajaxuploadFieldformat'] = 'Формат данных, сохранённых в `ajaxuploadFieldname`.';
-$_lang['ajaxupload.ajaxupload2formit.ajaxuploadFieldname'] = 'Поле Formit для сохранения (уже) загруженных имён/путей файлов.';
+$_lang['ajaxupload.ajaxupload2formit.ajaxuploadCacheExpires'] = 'Время истечения срока действия кэша AjaxUpload (в часах).';
+$_lang['ajaxupload.ajaxupload2formit.ajaxuploadFieldformat'] = 'Формат данных, сохранённых в `ajaxuploadUid`.';
 $_lang['ajaxupload.ajaxupload2formit.ajaxuploadSanitizeFilename'] = 'Дезинфекция имени загружаемого файла.';
-$_lang['ajaxupload.ajaxupload2formit.ajaxuploadTarget'] = 'Путь для сохранения (уже) загруженных файлов (относительно $modx->getOption[\'assetsPath\']). Папка должна существовать или должна быть создана, и она должна быть доступна для записи PHP.';
-$_lang['ajaxupload.ajaxupload2formit.ajaxuploadUid'] = 'Уникальный id очереди загрузки.';
+$_lang['ajaxupload.ajaxupload2formit.ajaxuploadTarget'] = 'Целевой путь для загружаемых файлов. Папка должна существовать или быть создана и должна быть доступна для записи PHP.';
+$_lang['ajaxupload.ajaxupload2formit.ajaxuploadTargetRelativePath'] = 'Путь, с которым сравнивается целевой путь загрузки. Пути загружаемых файлов обрабатываются FormIt относительно этого пути.';
+$_lang['ajaxupload.ajaxupload2formit.ajaxuploadUid'] = 'Список уникальных идентификаторов очередей загрузки, разделенных запятыми. Также имена полей, используемых в хуках.';
 $_lang['ajaxupload.ajaxupload2formit.csv'] = 'CSV';
 $_lang['ajaxupload.ajaxupload2formit.json'] = 'JSON';
-$_lang['ajaxupload.ajaxuploadrequired.ajaxuploadRequiredMessage'] = 'Сообщение об ошибке, добавляемое, когда файл не загружен. По умолчанию используется запись лексикона `ajaxupload.uploadRequired`.';
-$_lang['ajaxupload.formit2ajaxupload.ajaxuploadFieldformat'] = 'Формат данных, сохранённых в `ajaxuploadFieldname`.';
-$_lang['ajaxupload.formit2ajaxupload.ajaxuploadFieldname'] = 'Поле Formit для сохранения (уже) загруженных имён/путей файлов.';
-$_lang['ajaxupload.formit2ajaxupload.ajaxuploadTarget'] = 'Путь для сохранения (уже) загруженных файлов (относительно $modx->getOption[\'assetsPath\']).';
-$_lang['ajaxupload.formit2ajaxupload.ajaxuploadUid'] = 'Уникальный id очереди загрузки.';
+$_lang['ajaxupload.ajaxuploadattachments.ajaxuploadFieldformat'] = 'Формат данных, сохранённых в `ajaxuploadUid`.';
+$_lang['ajaxupload.ajaxuploadattachments.ajaxuploadTargetRelativePath'] = 'Путь, с которым сравнивается целевой путь загрузки. Пути загружаемых файлов обрабатываются FormIt относительно этого пути.';
+$_lang['ajaxupload.ajaxuploadattachments.ajaxuploadUid'] = 'Список уникальных идентификаторов очередей загрузки, разделенных запятыми. Также имена полей, используемых в хуках.';
+$_lang['ajaxupload.ajaxuploadremove.ajaxuploadFieldformat'] = 'Формат данных, сохраненных в `ajaxuploadUid`.';
+$_lang['ajaxupload.ajaxuploadremove.ajaxuploadTargetRelativePath'] = 'Путь, с которым сравнивается целевой путь загрузки. Пути загружаемых файлов обрабатываются FormIt относительно этого пути.';
+$_lang['ajaxupload.ajaxuploadremove.ajaxuploadUid'] = 'Список уникальных идентификаторов очереди загрузки, разделенных запятыми. Также имена полей, используемых в хуках.';
+$_lang['ajaxupload.ajaxuploadrequired.ajaxuploadFieldformat'] = 'Формат данных, сохраняемых в `ajaxuploadUid`.';
+$_lang['ajaxupload.ajaxuploadrequired.ajaxuploadRequiredMessage'] = 'Добавлено сообщение об ошибке, когда файл не загружен.';
+$_lang['ajaxupload.ajaxuploadrequired.ajaxuploadUid'] = 'Список уникальных идентификаторов очереди загрузки, разделенных запятыми. Также имена полей, используемых в хуках.';
+$_lang['ajaxupload.formit2ajaxupload.ajaxuploadCacheExpires'] = 'Время истечения кэша AjaxUpload (в часах).';
+$_lang['ajaxupload.formit2ajaxupload.ajaxuploadFieldformat'] = 'Формат данных, сохраняемых в `ajaxuploadUid`.';
+$_lang['ajaxupload.formit2ajaxupload.ajaxuploadTarget'] = 'Целевой путь, содержащий существующие файлы, подготовленные для FilePond javascript.';
+$_lang['ajaxupload.formit2ajaxupload.ajaxuploadTargetRelativePath'] = 'Путь, с которым сравнивается целевой путь загрузки. Пути загружаемых файлов обрабатываются FormIt относительно этого пути.';
+$_lang['ajaxupload.formit2ajaxupload.ajaxuploadUid'] = 'Список уникальных идентификаторов очереди загрузки, разделенных запятыми. Также имена полей, используемых в хуках.';
 $_lang['ajaxupload.formit2ajaxupload.csv'] = 'CSV';
 $_lang['ajaxupload.formit2ajaxupload.json'] = 'JSON';
+$_lang['ajaxupload.ajaxupload.targetRelativePath'] = 'Значение свойства ajaxuploadTargetRelativePath, используемое в соответствующем вызове фрагмента FormIt.';
+$_lang['ajaxupload.ajaxuploadremove.ajaxuploadTargetMediasource'] = 'Если не пустой, то целевой путь префиксен с базовым путем медиаисточника. В противном случае целевому пути присваивается путь к активам MODX.';
+$_lang['ajaxupload.ajaxupload.targetMediasource'] = 'Значение свойства ajaxuploadTargetMediasource, используемое в соответствующем вызове сниппета FormIt.';
+$_lang['ajaxupload.formit2ajaxupload.ajaxuploadTargetMediasource'] = 'Если не пустой, то целевой путь префиксен с базовым путем медиаисточника. В противном случае целевому пути присваивается путь к активам MODX.';
+$_lang['ajaxupload.ajaxupload2formit.ajaxuploadTargetMediasource'] = 'Если не пустой, то целевой путь префиксен с базовым путем медиаисточника. В противном случае целевому пути присваивается путь к активам MODX.';
+$_lang['ajaxupload.ajaxuploadattachments.ajaxuploadTargetMediasource'] = 'Если не пустой, то целевой путь префиксен с базовым путем медиаисточника. В противном случае целевому пути присваивается путь к активам MODX.';
